Add complete implementation of the 2D world MCP.

+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/monteCarloPlanning/RangeScanner.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 public class RangeScanner
 {
    private final int numPoints = 100;
-   private final float maxRange = 10.0f;
+   private final int maxRange = 10;
 
    public Point2D getClosestIntersectionPoint(Point2D start_point, Point2D end_point, Point2D obstacle_center, float obstacle_size_x, float obstacle_size_y)
    {
@@ -250,7 +250,7 @@ public class RangeScanner
       return numPoints;
    }
 
-   public float getMaxRange()
+   public int getMaxRange()
    {
       return maxRange;
    }
